+++ src/strategy/DailyPricing.java
@@ -1,5 +1,5 @@
 package strategy;
 
 public class DailyPricing implements PricingStrategy {
-    public double calculatePrice(int duration) { return duration * 60; }
+    public float calculatePrice(int duration) { return duration * 60; }
 }
