+++ src/strategy/PricingStrategy.java
@@ -1,5 +1,5 @@
 package strategy;
 
 public interface PricingStrategy {
-    double calculatePrice(int duration);
+    float calculatePrice(int duration);
 }
